"edit.h/.cc" splitted to limit rebuilding dependencies
