+++ rtengine/improcfun.cc
@@ -44,6 +44,7 @@
 #include "procparams.h"
 #include "../rtgui/ppversion.h"
 #include "../rtgui/guiutils.h"
+#include "../rtgui/editcallbacks.h"
 
 #undef CLIPD
 #define CLIPD(a) ((a)>0.0f?((a)<1.0f?(a):1.0f):0.0f)
