+++ rtgui/dirpyrdenoise.cc
@@ -21,10 +21,10 @@
 
 #include "dirpyrdenoise.h"
 
-#include "edit.h"
 #include "guiutils.h"
 
 #include "../rtengine/procparams.h"
+#include "editbuffer.h"
 
 using namespace rtengine;
 using namespace rtengine::procparams;
@@ -63,7 +63,7 @@ DirPyrDenoise::DirPyrDenoise () : FoldableToolPanel(this, "dirpyrdenoise", M("TP
     NoiscurveEditorG = new CurveEditorGroup (options.lastDenoiseCurvesDir, M("TP_DIRPYRDENOISE_LUMINANCE_CURVE"));
     //curveEditorG = new CurveEditorGroup (options.lastLabCurvesDir);
     NoiscurveEditorG->setCurveListener (this);
-    defaultCurve = rtengine::DirPyrDenoiseParams().lcurve;
+    defaultCurve = rtengine::procparams::DirPyrDenoiseParams().lcurve;
     lshape = static_cast<FlatCurveEditor*>(NoiscurveEditorG->addCurve(CT_Flat, "", nullptr, false, false));
     lshape->setIdentityValue(0.);
     lshape->setResetCurve(FlatCurveType(defaultCurve.at(0)), defaultCurve);
@@ -139,7 +139,7 @@ DirPyrDenoise::DirPyrDenoise () : FoldableToolPanel(this, "dirpyrdenoise", M("TP
 
     CCcurveEditorG = new CurveEditorGroup (options.lastDenoiseCurvesDir, M("TP_DIRPYRDENOISE_CHROMINANCE_CURVE"));
     CCcurveEditorG->setCurveListener (this);
-    defaultCurve = rtengine::DirPyrDenoiseParams().cccurve;
+    defaultCurve = rtengine::procparams::DirPyrDenoiseParams().cccurve;
     ccshape = static_cast<FlatCurveEditor*>(CCcurveEditorG->addCurve(CT_Flat, "", nullptr, false, false));
     ccshape->setIdentityValue(0.);
     ccshape->setResetCurve(FlatCurveType(defaultCurve.at(0)), defaultCurve);
