+++ rtgui/labgrid.cc
@@ -38,6 +38,8 @@
 
 #include "labgrid.h"
 
+#include "options.h"
+
 using rtengine::Color;
