+++ rtgui/mydiagonalcurve.cc
@@ -18,6 +18,7 @@
  */
 #include "mydiagonalcurve.h"
 #include "../rtengine/curves.h"
+#include "editcallbacks.h"
 #include <cstring>
 #include <gdkmm/types.h>
