+++ rtgui/preprocess.cc
@@ -21,6 +21,7 @@
 #include "preprocess.h"
 
 #include "guiutils.h"
+#include "options.h"
 
 #include "../rtengine/procparams.h"
