+++ rtgui/rawcacorrection.cc
@@ -21,6 +21,7 @@
 #include "eventmapper.h"
 #include "guiutils.h"
 #include "rtimage.h"
+#include "options.h"
 
 #include "../rtengine/procparams.h"
