+++ rtgui/rawexposure.cc
@@ -21,6 +21,7 @@
 #include "rawexposure.h"
 
 #include "guiutils.h"
+#include "options.h"
 
 #include "../rtengine/procparams.h"
