+++ rtgui/sharpenedge.cc
@@ -22,6 +22,7 @@
 #include "sharpenedge.h"
 
 #include "guiutils.h"
+#include "options.h"
 
 #include "../rtengine/procparams.h"
