+++ rtgui/toolpanel.h
@@ -22,10 +22,10 @@
 #include <gtkmm.h>
 #include <glibmm.h>
 #include "../rtengine/rtengine.h"
+#include "editbuffer.h"
 #include "guiutils.h"
 #include "multilangmgr.h"
 #include "paramsedited.h"
-#include "edit.h"
 
 class ToolPanel;
 class FoldableToolPanel;
