+++ rtgui/xtransrawexposure.cc
@@ -21,6 +21,7 @@
 #include "xtransrawexposure.h"
 
 #include "guiutils.h"
+#include "options.h"
 
 #include "../rtengine/procparams.h"
